Improve the dot decoration placement as we edit
The dot decorations would often appear in the wrong spot as we edit the
document because they updated after every change using locations from
the last results.

This commit stops updating the test result decorations when there are
unsaved changes to the document. Using the `onDidChangeTextDocument`
event handler, we can detect when the document is saved or updated by an
external editor and update accordingly.

To keep the decorations in place as we edit or format the document, the
the decoration types are now created with a `DecorationRangeBehavior`
that prevents the range from growing.

This commit also changes the way the inline error messages are cached to
keep the error messages visible in all instances of a document, not just
in the active editor.